more cleanup of git commits
