+++ firmware/src/core/utils/TaskManager.cpp
@@ -1,17 +1,22 @@
-#include "TaskManager.h"
-#include "GlobalResources.h"
+#include "HTTPClientWrapper.h"
 #include "Utils.h"
 #include <HTTPClient.h>
-#include <memory>
 
-TaskManager *TaskManager::instance = nullptr;
-QueueHandle_t TaskManager::requestQueue = nullptr;
-QueueHandle_t TaskManager::responseQueue = nullptr;
-volatile uint32_t TaskManager::activeRequests = 0;
-volatile uint32_t TaskManager::maxConcurrentRequests = 0;
-int TaskManager::taskParamsCount = 0;
+HTTPClientWrapper *HTTPClientWrapper::instance = nullptr;
+SemaphoreHandle_t HTTPClientWrapper::requestSemaphore = nullptr;
+QueueHandle_t HTTPClientWrapper::requestQueue = nullptr;
+QueueHandle_t HTTPClientWrapper::responseQueue = nullptr;
 
-TaskManager::TaskManager() {
+// Initialize static debug counters
+volatile uint32_t HTTPClientWrapper::activeRequests = 0;
+volatile uint32_t HTTPClientWrapper::maxConcurrentRequests = 0;
+
+HTTPClientWrapper::HTTPClientWrapper() {
+    if (!requestSemaphore) {
+        requestSemaphore = xSemaphoreCreateBinary();
+        Utils::setBusy(false);
+        xSemaphoreGive(requestSemaphore);
+    }
     if (!requestQueue) {
         requestQueue = xQueueCreate(REQUEST_QUEUE_SIZE, REQUEST_QUEUE_ITEM_SIZE);
     }
@@ -20,114 +25,146 @@ TaskManager::TaskManager() {
     }
 }
 
-TaskManager *TaskManager::getInstance() {
+HTTPClientWrapper *HTTPClientWrapper::getInstance() {
     if (!instance) {
-        instance = new TaskManager();
+        instance = new HTTPClientWrapper();
     }
     return instance;
 }
 
-bool TaskManager::addTask(std::unique_ptr<Task> task) {
-    if (isUrlInQueue(task->url)) {
+bool HTTPClientWrapper::addRequest(const String &url, ResponseCallback callback, PreProcessCallback preProcessResponse) {
+    if (isUrlInQueue(url)) {
+        Serial.printf("Request already in queue: %s\n", url.c_str());
         return false;
     }
 
-    auto *params = new TaskParams{task->url, task->callback, task->preProcessResponse, task->taskExec};
-    taskParamsCount++; // Increment the count
-#ifdef TASKMANAGER_DEBUG
-    Serial.printf("TaskParams created: %d\n", taskParamsCount);
-#endif
+    auto *params = new RequestParams{url, callback, preProcessResponse};
 
     if (xQueueSend(requestQueue, &params, 0) != pdPASS) {
         delete params;
-        taskParamsCount--;
-#ifdef TASKMANAGER_DEBUG
-        Serial.printf("TaskParams deleted (queue full): %d\n", taskParamsCount);
-#endif
+        Serial.println("Failed to queue HTTP request");
         return false;
     }
 
+    Serial.printf("Queued request: %s (Queue length: %d)\n", url.c_str(), uxQueueMessagesWaiting(requestQueue));
     return true;
 }
 
-void TaskManager::processAwaitingTasks() {
+void HTTPClientWrapper::processRequestQueue() {
     // First check if there are any requests to process
     if (uxQueueMessagesWaiting(requestQueue) == 0) {
         return; // No requests in queue
     }
 
-    if (xSemaphoreTake(taskSemaphore, 0) != pdTRUE) {
+    // Only try to take semaphore if we have work to do
+    if (xSemaphoreTake(requestSemaphore, 0) != pdTRUE) {
+        // Serial.println("⚠️ Semaphore blocked - request already in progress");
         return;
     }
 
     Utils::setBusy(true);
+    Serial.println("✅ Obtained semaphore");
     activeRequests++;
-
     if (activeRequests > maxConcurrentRequests) {
         maxConcurrentRequests = activeRequests;
     }
-#ifdef TASKMANAGER_DEBUG
-    Serial.println("✅ Obtained semaphore");
     Serial.printf("Active requests: %d (Max seen: %d)\n", activeRequests, maxConcurrentRequests);
-#endif
 
     // Get next request
-    TaskParams *taskParams = nullptr;
-    if (xQueueReceive(requestQueue, &taskParams, 0) != pdPASS) {
+    RequestParams *requestParams;
+    if (xQueueReceive(requestQueue, &requestParams, 0) != pdPASS) {
         Serial.println("⚠️ Queue empty after size check!");
         activeRequests--;
         Utils::setBusy(false);
-        xSemaphoreGive(taskSemaphore);
+        xSemaphoreGive(requestSemaphore);
         return;
     }
 
-#ifdef TASKMANAGER_DEBUG
     Serial.printf("Processing request: %s (Remaining in queue: %d)\n",
-                  taskParams->url.c_str(),
+                  requestParams->url.c_str(),
                   uxQueueMessagesWaiting(requestQueue));
-#endif
 
+    // Create task to handle request
     TaskHandle_t taskHandle;
     BaseType_t result = xTaskCreate(
-        [](void *params) {
-            auto *taskParams = static_cast<TaskParams *>(params);
-            taskParams->taskExec();
-            delete taskParams; // Ensure cleanup after execution
-            taskParamsCount--; // Decrement the count
-            // Serial.printf("TaskParams deleted: %d\n", taskParamsCount);
-
-            Utils::setBusy(false);
-            xSemaphoreGive(taskSemaphore);
-            Serial.println("✅ Released semaphore");
-            vTaskDelete(nullptr);
-        },
-        "TASK_EXEC",
+        httpTask,
+        "HTTP_REQ",
         STACK_SIZE,
-        taskParams,
+        requestParams, // Pass params to task
         TASK_PRIORITY,
         &taskHandle);
 
     if (result != pdPASS) {
         Serial.println("Failed to create HTTP request task");
-        delete taskParams; // Ensure the object is deleted if task creation fails
-        taskParamsCount--; // Decrement the count if task creation fails
-        Serial.printf("TaskParams deleted (task creation failed): %d\n", taskParamsCount);
+        delete requestParams;
         Utils::setBusy(false);
-        xSemaphoreGive(taskSemaphore);
+        xSemaphoreGive(requestSemaphore);
     }
 }
 
-void TaskManager::processTaskResponses() {
+void HTTPClientWrapper::httpTask(void *params) {
+    auto *requestParams = static_cast<RequestParams *>(params);
 
-#ifdef TASKMANAGER_DEBUG
-    // Periodically check for TaskParams leaks
-    static unsigned long lastLeakCheck = 0;
-    if (millis() - lastLeakCheck > 30000) { // Check every 30 seconds
-        TaskManager::checkForLeaks();
-        lastLeakCheck = millis();
+    Serial.printf("🔵 Starting HTTP request for: %s\n", requestParams->url.c_str());
+
+    {
+        HTTPClient http;
+        WiFiClientSecure client;
+        client.setInsecure();
+
+        http.begin(client, requestParams->url);
+        http.setTimeout(10000); // 10 second timeout
+
+        int httpCode = http.GET();
+        String response;
+
+        if (httpCode > 0) {
+            response = http.getString();
+        }
+
+        http.end();
+        client.stop();
+
+        // Explicitly reset the objects
+        http.~HTTPClient(); // Call the destructor
+        new (&http) HTTPClient(); // Reinitialize using placement new
+
+        client.~WiFiClientSecure(); // Call the destructor
+        new (&client) WiFiClientSecure(); // Reinitialize using placement new
+
+        // Call preProcessResponse if provided
+        if (requestParams->preProcessResponse) {
+            requestParams->preProcessResponse(httpCode, response);
+        }
+
+        auto *responseData = new ResponseData{
+            httpCode,
+            response,
+            requestParams->callback};
+
+        if (xQueueSend(responseQueue, &responseData, 0) != pdPASS) {
+            Serial.println("Failed to queue response");
+            delete responseData;
+        }
     }
-#endif
 
+    // UBaseType_t highWater = uxTaskGetStackHighWaterMark(NULL);
+    // Serial.print("Remaining task stack space: ");
+    // Serial.println(highWater);
+
+    Serial.printf("🟢 Completed HTTP request for: %s\n", requestParams->url.c_str());
+    activeRequests--;
+    Serial.printf("Active requests now: %d\n", activeRequests);
+
+    delete requestParams;
+    Utils::setBusy(false);
+    xSemaphoreGive(requestSemaphore);
+    Serial.println("✅ Released semaphore");
+    vTaskDelete(nullptr);
+}
+
+void HTTPClientWrapper::processResponseQueue() {
+    // Early exit if queue is empty
     if (uxQueueMessagesWaiting(responseQueue) == 0) {
         return;
     }
@@ -135,21 +172,23 @@ void TaskManager::processTaskResponses() {
     ResponseData *responseData;
     while (xQueueReceive(responseQueue, &responseData, 0) == pdPASS) {
         responseData->callback(responseData->httpCode, responseData->response);
-        delete responseData; // Ensure the object is deleted after processing
+        delete responseData;
     }
 }
 
-bool TaskManager::isUrlInQueue(const String &url) {
+bool HTTPClientWrapper::isUrlInQueue(const String &url) {
+    // Iterate through the queue to check for duplicate URLs
     UBaseType_t queueLength = uxQueueMessagesWaiting(requestQueue);
     for (UBaseType_t i = 0; i < queueLength; i++) {
-        TaskParams *taskParams;
-        if (xQueuePeek(requestQueue, &taskParams, 0) == pdPASS) {
-            if (taskParams->url == url) {
-                return true;
+        RequestParams *requestParams;
+        if (xQueuePeek(requestQueue, &requestParams, 0) == pdPASS) {
+            if (requestParams->url == url) {
+                return true; // Duplicate URL found
             }
-            xQueueReceive(requestQueue, &taskParams, 0);
-            xQueueSend(requestQueue, &taskParams, 0);
+            // Move to the next item in the queue
+            xQueueReceive(requestQueue, &requestParams, 0);
+            xQueueSend(requestQueue, &requestParams, 0);
         }
     }
-    return false;
+    return false; // No duplicate URL found
 }
